+++ src/linker/pipeline.py
@@ -1,7 +1,10 @@
+from collections import defaultdict
 from pathlib import Path
-from typing import Callable, Optional
+from typing import Any, Callable, Dict, Optional, Tuple
 
 from linker.configuration import Config
+from linker.implementation import Implementation
+from linker.utilities.general_utils import exit_with_validation_error
 
 
 class Pipeline:
@@ -9,7 +12,46 @@ class Pipeline:
 
     def __init__(self, config: Config):
         self.config = config
-        self.implementations = config.implementations
+        self.implementations = self._get_implementations()
+        self._validate()
+
+    def get_implementation_name(self, step_name: str) -> str:
+        return self.config.pipeline["steps"][step_name]["implementation"]["name"]
+
+    def get_implementation_config(self, step_name: str) -> Optional[Dict[str, Any]]:
+        return self.config.pipeline["steps"][step_name]["implementation"].get(
+            "configuration", None
+        )
+
+    def _get_implementations(self) -> Tuple[Implementation, ...]:
+        resources = {key: self.config.environment.get(key) for key in ["slurm", "spark"]}
+        return tuple(
+            Implementation(
+                step=step,
+                implementation_name=self.get_implementation_name(step.name),
+                implementation_config=self.get_implementation_config(step.name),
+                container_engine=self.config.container_engine,
+                resources=resources,
+            )
+            for step in self.config.schema.steps
+        )
+
+    def _validate(self) -> None:
+        """Validates the pipeline."""
+
+        errors = {**self._validate_implementations()}
+
+        if errors:
+            exit_with_validation_error(errors)
+
+    def _validate_implementations(self) -> Dict:
+        """Validates each individual Implementation instance."""
+        errors = defaultdict(dict)
+        for implementation in self.implementations:
+            implementation_errors = implementation.validate()
+            if implementation_errors:
+                errors["IMPLEMENTATION ERRORS"][implementation.name] = implementation_errors
+        return errors
 
     def run(
         self,
